+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.0] - 2025-07-19
+
 ### Added
 
 - Puzzle preview and navigation buttons added above freeplay puzzle select list.